refactor: reallocate scheme type to a proper isolate context
This prevents name collisions the exposed constants and functions and keeps the logic isolate according to the context of scheme
