+++ resolver_factory.go
@@ -1,16 +1,19 @@
 package core
 
-import "errors"
+import (
+	"errors"
+	"github.com/lot-sh/core/scheme"
+)
 
 // ResolverFactory returns the implementation of Resolver
 // which can handle the SchemeType passed as argument
 //
 // • when there is not a Resolver associated to the SchemeType
 // passed as argument it will return a nil Resolver and an error
-func ResolverFactory(st SchemeType) (Resolver, error) {
+func ResolverFactory(st scheme.SchemeType) (Resolver, error) {
 	var resolver Resolver
 	switch st {
-	case HTTP, HTTPS:
+	case scheme.HTTP, scheme.HTTPS:
 		resolver = &HTTPResolver{}
 		return resolver, nil
 	}
