+++ resource_factory.go
@@ -1,15 +1,18 @@
 package core
 
-import "errors"
+import (
+	"errors"
+	"github.com/lot-sh/core/scheme"
+)
 
 // ResourceFactory function returns a Resource given a locator
 func ResourceFactory(locator string) (Resource, error) {
-	var scheme SchemeType = GetSchemeTypeFrom(locator)
+	var sch scheme.SchemeType = scheme.GetSchemeTypeFrom(locator)
 	var resource Resource = Resource{
 		locator,
-		scheme,
+		sch,
 	}
-	if scheme == UNKNOWN {
+	if sch == scheme.UNKNOWN {
 		return resource, errors.New("Unknown scheme, the locator may be a invalid one")
 	}
 	return resource, nil
