+++ resource_factory_test.go
@@ -1,6 +1,9 @@
 package core
 
-import "testing"
+import (
+	"testing"
+	"github.com/lot-sh/core/scheme"
+)
 
 func TestResourceFactoryShouldWorkAsExpected(t *testing.T) {
 	locator := "lot:QmT5NvUtoM5nWFfrQdVrFtvGfKFmG7AHE8P34isapyhCxX"
@@ -8,8 +11,8 @@ func TestResourceFactoryShouldWorkAsExpected(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	if resource.Scheme != LOT {
-		t.Errorf("the scheme of %s should be %s, found %s", locator, LOT, resource.Scheme)
+	if resource.Scheme != scheme.LOT {
+		t.Errorf("the scheme of %s should be %s, found %s", locator, scheme.LOT, resource.Scheme)
 	}
 	if resource.Locator != locator {
 		t.Errorf("the locator should be %s, found %s", locator, resource.Locator)
